# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invillia Test",
        "description": "Invillia Test",
        "contact": {
            "email": "jonathanbeltrao@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get all Customers and its info",
                "description": "Get all Customers and its info",
                "operationId": "getCustomers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Customer ID",
                                            "type": "string"
                                        },
                                        "name": {
                                            "description": "Customer Name",
                                            "type": "string"
                                        },
                                        "orders": {
                                            "description": "Customer Orders",
                                            "type": "string"
                                        },
                                        "phones": {
                                            "description": "Customer Phones",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "id": 1,
                                            "customer_id": 1,
                                            "address": {
                                                "id": 10,
                                                "order_id": 1,
                                                "name": "Name 1",
                                                "address": "Address 1",
                                                "city": "City 1",
                                                "country": "Country 1"
                                            },
                                            "items": [
                                                {
                                                    "id": 9,
                                                    "order_id": 1,
                                                    "title": "Title 1",
                                                    "note": "Note 1",
                                                    "quantity": 745,
                                                    "price": 123.45
                                                }
                                            ],
                                            "phones": [
                                                {
                                                    "id": 76,
                                                    "customer_id": 1,
                                                    "number": "2345678",
                                                    "created_at": "2020-12-08T05:45:15.000000Z",
                                                    "updated_at": "2020-12-08T05:45:15.000000Z",
                                                    "deleted_at": null
                                                },
                                                {
                                                    "id": 77,
                                                    "customer_id": 1,
                                                    "number": "1234567",
                                                    "created_at": "2020-12-08T05:45:15.000000Z",
                                                    "updated_at": "2020-12-08T05:45:15.000000Z",
                                                    "deleted_at": null
                                                }
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/customer/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get a Customers by its ID",
                "description": "Get a Customer and its info",
                "operationId": "getCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Customer ID",
                                            "type": "string"
                                        },
                                        "name": {
                                            "description": "Customer Name",
                                            "type": "string"
                                        },
                                        "orders": {
                                            "description": "Customer Orders",
                                            "type": "string"
                                        },
                                        "phones": {
                                            "description": "Customer Phones",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 1,
                                        "customer_id": 1,
                                        "address": {
                                            "id": 10,
                                            "order_id": 1,
                                            "name": "Name 1",
                                            "address": "Address 1",
                                            "city": "City 1",
                                            "country": "Country 1"
                                        },
                                        "items": [
                                            {
                                                "id": 9,
                                                "order_id": 1,
                                                "title": "Title 1",
                                                "note": "Note 1",
                                                "quantity": 745,
                                                "price": 123.45
                                            }
                                        ],
                                        "phones": [
                                            {
                                                "id": 76,
                                                "customer_id": 1,
                                                "number": "2345678",
                                                "created_at": "2020-12-08T05:45:15.000000Z",
                                                "updated_at": "2020-12-08T05:45:15.000000Z",
                                                "deleted_at": null
                                            },
                                            {
                                                "id": 77,
                                                "customer_id": 1,
                                                "number": "1234567",
                                                "created_at": "2020-12-08T05:45:15.000000Z",
                                                "updated_at": "2020-12-08T05:45:15.000000Z",
                                                "deleted_at": null
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}